add a controller reset step to election ritual
The controller stores a lot of state across runs, so that if it crashes
it can be restarted without nullifying the election. That state needs to
be reset when a new election is started (e.g. you wouldn't want
encrypted ballots from the previous election persisting into the new
election). This commit adds a way to do that, and reminds you to do this
step during the election setup.